lock für exportierte Bestellungen eingerichtet

+++ admin/ajax/orders_delete.php
@@ -8,8 +8,17 @@ $year = strtok(".");
 $orderDate = $year."-".$month."-".$day;
 
 $db = new db_connection();
-$result = $db->deleteData("orders", 
-"idProduct=".$_POST["productId"]." AND hook=".$_POST["orderHook"]." AND idCustomer=".$_POST["customer"]." AND orderDate='".$orderDate."'");
 
-echo $result;
+$locked = $db->getData("orders",array('locked'),
+    "idProduct=".$_POST["productId"]." AND hook=".$_POST["orderHook"]." AND idCustomer=".$_POST["customer"]." AND orderDate='".$orderDate."'")[0]['locked'];
+
+if(!$locked) {
+    $result = $db->deleteData("orders",
+        "idProduct=" . $_POST["productId"] . " AND hook=" . $_POST["orderHook"] . " AND idCustomer=" . $_POST["customer"] . " AND orderDate='" . $orderDate . "'");
+}
+else {
+    $result = "Der Artikel kann nicht gelöscht werden, er wurde schon exportiert.";
+}
+$output = json_encode($result);
+return $output;
 ?>
